+++ components/camel-couchdb/src/test/java/org/apache/camel/component/couchdb/integration/CouchDbCrudIT.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.couchdb;
+package org.apache.camel.component.couchdb.integration;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -26,6 +26,8 @@ import org.apache.camel.Endpoint;
 import org.apache.camel.EndpointInject;
 import org.apache.camel.Exchange;
 import org.apache.camel.builder.RouteBuilder;
+import org.apache.camel.component.couchdb.CouchDbConstants;
+import org.apache.camel.component.couchdb.CouchDbOperations;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.junit.jupiter.api.MethodOrderer.OrderAnnotation;
 import org.junit.jupiter.api.Order;
@@ -40,7 +42,7 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 @TestMethodOrder(OrderAnnotation.class)
 @TestInstance(Lifecycle.PER_CLASS)
-public class CouchDbCrudTest extends CouchDbTestSupport {
+public class CouchDbCrudIT extends CouchDbTestSupport {
 
     @EndpointInject("mock:deleteNotifications")
     private MockEndpoint mockDeleteNotifications;
