+++ components/camel-couchdb/src/test/java/org/apache/camel/component/couchdb/integration/CouchDbTestSupport.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.couchdb;
+package org.apache.camel.component.couchdb.integration;
 
 import org.apache.camel.test.infra.couchdb.services.CouchDbService;
 import org.apache.camel.test.infra.couchdb.services.CouchDbServiceFactory;
